requirements-dbt.txt
dbt-trino==1.9.3
dbt-core==1.10.10
